fix violations of Sonarqube rule java:S2184

+++ src/main/java/org/apache/commons/lang3/text/StrBuilder.java
@@ -162,7 +162,7 @@ public class StrBuilder implements CharSequence, Appendable, Serializable, Build
         @Override
         public long skip(long n) {
             if (pos + n > StrBuilder.this.size()) {
-                n = StrBuilder.this.size() - pos;
+                n = (long) StrBuilder.this.size() - pos;
             }
             if (n < 0) {
                 return 0;
