+++ src/main/java/org/apache/commons/lang3/time/DateUtils.java
@@ -403,10 +403,10 @@ public class DateUtils {
         // Fragments bigger than a day require a breakdown to days
         switch (fragment) {
             case Calendar.YEAR:
-                result += unit.convert(calendar.get(Calendar.DAY_OF_YEAR) - offset, TimeUnit.DAYS);
+                result += unit.convert((long) calendar.get(Calendar.DAY_OF_YEAR) - offset, TimeUnit.DAYS);
                 break;
             case Calendar.MONTH:
-                result += unit.convert(calendar.get(Calendar.DAY_OF_MONTH) - offset, TimeUnit.DAYS);
+                result += unit.convert((long) calendar.get(Calendar.DAY_OF_MONTH) - offset, TimeUnit.DAYS);
                 break;
             default:
                 break;
